Splitting pmodel into submodules
