+++ pyrealm/pmodel/__init__.py
@@ -1,1 +1,43 @@
-"""The :mod:`~pyrealm.pmodel` provides the main PModel class + extensions to do..."""
+"""The :mod:`~pyrealm.pmodel` module includes four submodules:
+
+* The :mod:`~pyrealm.pmodel.pmodel` submodule provides classes implementing the core
+  calculations of the P  Model.
+* The :mod:`~pyrealm.pmodel.functions` submodule provides key standalone functions for
+  those calculations.
+* The :mod:`~pyrealm.pmodel.isotopes` submodule provides a class to estimate isotopic
+  discrimination within the P Model.
+* The :mod:`~pyrealm.pmodel.competition` submodule provides a competition model for the
+  C3 and C4 photosynthetic pathways.
+
+Note that the documentation of functions and methods includes two lists of parameters:
+
+Parameters
+  These are the arguments specific to the class, method or function signature.
+
+PModel Parameters
+  These are shared parameters of the PModel used by this function, which are taken from
+  a :mod:`~pyrealm.params.PModelParams` instance. These can be changed by the user but
+  are typically used to configure an entire analysis rather than a single function.
+"""  # noqa: D210, D415
+
+# This __init__ file imports the following members from submodules in order to
+# flatten the namespace for the main public components and setup.cfg applies
+# # noqa: F401 to the whole file.
+
+from pyrealm.pmodel.competition import C3C4Competition
+from pyrealm.pmodel.functions import (
+    calc_co2_to_ca,
+    calc_density_h2o,
+    calc_ftemp_arrh,
+    calc_ftemp_inst_rd,
+    calc_ftemp_inst_vcmax,
+    calc_ftemp_kphio,
+    calc_gammastar,
+    calc_kmm,
+    calc_ns_star,
+    calc_patm,
+    calc_soilmstress,
+    calc_viscosity_h2o,
+)
+from pyrealm.pmodel.isotopes import CalcCarbonIsotopes
+from pyrealm.pmodel.pmodel import PModel, PModelEnvironment
