+++ pyrealm/pmodel/pmodel.py
@@ -1,23 +1,14 @@
-"""This page contains the detailed documentation of the functions and classes in the
-:mod:`~pyrealm.pmodel` module. The documentation describes the key equations used in
-each member. In particular, note that the documentation of functions and methods
-includes two lists of parameters:
+"""The :mod:`~pyrealm.pmodel.pmodel` submodule provides the core implementation of the
+following core classes:
 
-Parameters
-  These are the arguments specific to the class, method or function signature.
-
-PModel Parameters
-  These are shared parameters of the PModel used by this function, which are taken from
-  a :mod:`~pyrealm.params.PModelParams` instance. These can be changed by the user but 
-  are typically used to configure an entire analysis rather than a single function.
-
-This module provides:
-
-* The :mod:`~pyrealm.pmodel.PModel` class, implementing an optimality based model of
-  photosynthesis, along with a set of functions to convert environmental forcings to key
-  parameters used within the model.
-
-The module also provides code for...  TODO: Update this.
+* PModelEnvironment:
+    Calculates the photosynthetic environment for locations.
+* PModel:
+    Applies the PModel to locations
+* CalcOptimalChi:
+    Estimates the optimal chi for locations, given an estimation method and settings
+* JMaxLimitation:
+    Estimates the Jmax limitation, given a method and settings
 """  # noqa D210, D415
 
 from typing import Optional, Union
@@ -27,737 +18,17 @@ import numpy as np
 from numpy.typing import NDArray
 
 from pyrealm import ExperimentalFeatureWarning
-from pyrealm.param_classes import C3C4Params, IsotopesParams, PModelParams
+from pyrealm.param_classes import C3C4Params, PModelParams
+from pyrealm.pmodel.functions import (
+    calc_co2_to_ca,
+    calc_ftemp_inst_rd,
+    calc_ftemp_inst_vcmax,
+    calc_ftemp_kphio,
+    calc_gammastar,
+    calc_kmm,
+    calc_ns_star,
+)
 from pyrealm.utilities import bounds_checker, check_input_shapes, summarize_attrs
-
-# TODO - Note that the typing currently does not enforce the dtype of ndarrays
-#        but it looks like the upcoming np.typing module might do this.
-
-
-def calc_density_h2o(
-    tc: NDArray,
-    patm: NDArray,
-    pmodel_params: PModelParams = PModelParams(),
-    safe: bool = True,
-) -> NDArray:
-    """Calculate water density.
-
-    Calculates the density of water as a function of temperature and atmospheric
-    pressure, using the Tumlirz Equation and coefficients calculated by
-    :cite:`Fisher:1975tm`.
-
-    Args:
-        tc: air temperature, °C
-        patm: atmospheric pressure, Pa
-        pmodel_params: An instance of :class:`~pyrealm.param_classes.PModelParams`.
-        safe: Prevents the function from estimating density below -30°C, where the
-            function behaves poorly
-
-    PModel Parameters:
-        lambda_: polynomial coefficients of Tumlirz equation (``fisher_dial_lambda``).
-        Po: polynomial coefficients of Tumlirz equation (``fisher_dial_Po``).
-        Vinf: polynomial coefficients of Tumlirz equation (``fisher_dial_Vinf``).
-
-    Returns:
-        Water density as a float in (g cm^-3)
-
-    Raises:
-        ValueError: if ``tc`` is less than -30°C and ``safe`` is True, or if the inputs
-            have incompatible shapes.
-
-    Examples:
-        >>> round(calc_density_h2o(20, 101325), 3)
-        998.206
-    """
-
-    # It doesn't make sense to use this function for tc < 0, but in particular
-    # the calculation shows wild numeric instability between -44 and -46 that
-    # leads to numerous downstream issues - see the extreme values documentation.
-    if safe and np.nanmin(tc) < -30:
-        raise ValueError(
-            "Water density calculations below about -30°C are "
-            "unstable. See argument safe to calc_density_h2o"
-        )
-
-    # Check input shapes, shape not used
-    _ = check_input_shapes(tc, patm)
-
-    # Get powers of tc, including tc^0 = 1 for constant terms
-    tc_pow = np.power.outer(tc, np.arange(0, 10))
-
-    # Calculate lambda, (bar cm^3)/g:
-    lambda_val = np.sum(pmodel_params.fisher_dial_lambda * tc_pow[..., :5], axis=-1)
-
-    # Calculate po, bar
-    po_val = np.sum(pmodel_params.fisher_dial_Po * tc_pow[..., :5], axis=-1)
-
-    # Calculate vinf, cm^3/g
-    vinf_val = np.sum(pmodel_params.fisher_dial_Vinf * tc_pow, axis=-1)
-
-    # Convert pressure to bars (1 bar <- 100000 Pa)
-    pbar = 1e-5 * patm
-
-    # Calculate the specific volume (cm^3 g^-1):
-    spec_vol = vinf_val + lambda_val / (po_val + pbar)
-
-    # Convert to density (g cm^-3) -> 1000 g/kg; 1000000 cm^3/m^3 -> kg/m^3:
-    rho = 1e3 / spec_vol
-
-    return rho
-
-
-def calc_ftemp_arrh(
-    tk: NDArray,
-    ha: float,
-    pmodel_params: PModelParams = PModelParams(),
-) -> NDArray:
-    r"""Calculate enzyme kinetics scaling factor.
-
-    Calculates the temperature-scaling factor :math:`f` for enzyme kinetics following an
-    Arrhenius response for a given temperature (``tk``, :math:`T`) and activation energy
-    (``ha``, :math:`H_a`).
-
-    Arrhenius kinetics are described as:
-
-    .. math::
-
-        x(T) = \exp(c - H_a / (T R))
-
-    The temperature-correction function :math:`f(T, H_a)` is:
-
-      .. math::
-        :nowrap:
-
-        \[
-            \begin{align*}
-                f &= \frac{x(T)}{x(T_0)} \\
-                  &= \exp \left( \frac{ H_a (T - T_0)}{T_0 R T}\right)
-                        \text{, or equivalently}\\
-                  &= \exp \left( \frac{ H_a}{R} \cdot
-                        \left(\frac{1}{T_0} - \frac{1}{T}\right)\right)
-            \end{align*}
-        \]
-
-    Args:
-        tk: Temperature (in Kelvin)
-        ha: Activation energy (in :math:`J \text{mol}^{-1}`)
-        pmodel_params: An instance of :class:`~pyrealm.param_classes.PModelParams`.
-
-    PModel Parameters:
-        To: a standard reference temperature (:math:`T_0`, ``k_To``)
-        R: the universal gas constant (:math:`R`, ``k_R``)
-
-    Returns:
-        Estimated float values for :math:`f`
-
-    Examples:
-        >>> # Relative rate change from 25 to 10 degrees Celsius (percent change)
-        >>> round((1.0-calc_ftemp_arrh( 283.15, 100000)) * 100, 4)
-        88.1991
-    """
-
-    # Note that the following forms are equivalent:
-    # exp( ha * (tk - 298.15) / (298.15 * kR * tk) )
-    # exp( ha * (tc - 25.0)/(298.15 * kR * (tc + 273.15)) )
-    # exp( (ha/kR) * (1/298.15 - 1/tk) )
-
-    tkref = pmodel_params.k_To + pmodel_params.k_CtoK
-
-    return np.exp(ha * (tk - tkref) / (tkref * pmodel_params.k_R * tk))
-
-
-def calc_ftemp_inst_rd(
-    tc: NDArray, pmodel_params: PModelParams = PModelParams()
-) -> NDArray:
-    r"""Calculate temperature scaling of dark respiration.
-
-    Calculates the temperature-scaling factor for dark respiration at a given
-    temperature (``tc``, :math:`T` in °C), relative to the standard reference
-    temperature :math:`T_o`, given the parameterisation in :cite:`Heskel:2016fg`.
-
-    .. math::
-
-            fr = \exp( b (T_o - T) -  c ( T_o^2 - T^2 ))
-
-    Args:
-        tc: Temperature (degrees Celsius)
-
-    PModel Parameters:
-        To: standard reference temperature (:math:`T_o`, ``k_To``)
-        b: empirically derived global mean coefficient
-            (:math:`b`, ``heskel_b``)
-        c: empirically derived global mean coefficient
-            (:math:`c`, ``heskel_c``)
-
-    Returns:
-        Values for :math:`fr`
-
-    Examples:
-        >>> # Relative percentage instantaneous change in Rd going from 10 to 25 degrees
-        >>> val = (calc_ftemp_inst_rd(25) / calc_ftemp_inst_rd(10) - 1) * 100
-        >>> round(val, 4)
-        250.9593
-    """
-
-    return np.exp(
-        pmodel_params.heskel_b * (tc - pmodel_params.k_To)
-        - pmodel_params.heskel_c * (tc**2 - pmodel_params.k_To**2)
-    )
-
-
-def calc_ftemp_inst_vcmax(
-    tc: NDArray, pmodel_params: PModelParams = PModelParams()
-) -> NDArray:
-    r"""Calculate temperature scaling of :math:`V_{cmax}`.
-
-    This function calculates the temperature-scaling factor :math:`f` of the
-    instantaneous temperature response of :math:`V_{cmax}`, given the temperature
-    (:math:`T`) relative to the standard reference temperature (:math:`T_0`), following
-    modified Arrhenius kinetics.
-
-    .. math::
-
-       V = f V_{ref}
-
-    The value of :math:`f` is given by :cite:`Kattge:2007db` (Eqn 1) as:
-
-    .. math::
-
-        f = g(T, H_a) \cdot
-                \frac{1 + \exp( (T_0 \Delta S - H_d) / (T_0 R))}
-                     {1 + \exp( (T \Delta S - H_d) / (T R))}
-
-    where :math:`g(T, H_a)` is a regular Arrhenius-type temperature response function
-    (see :func:`~pyrealm.pmodel.calc_ftemp_arrh`). The term :math:`\Delta S` is the
-    entropy factor, calculated as a linear function of :math:`T` in °C following
-    :cite:`Kattge:2007db` (Table 3, Eqn 4):
-
-    .. math::
-
-        \Delta S = a + b T
-
-    Args:
-        tc:  temperature, or in general the temperature relevant for
-            photosynthesis (°C)
-        pmodel_params: An instance of :class:`~pyrealm.param_classes.PModelParams`.
-
-    PModel Parameters:
-        Ha: activation energy (:math:`H_a`, ``kattge_knorr_Ha``)
-        Hd: deactivation energy (:math:`H_d`, ``kattge_knorr_Hd``)
-        To: standard reference temperature expressed in Kelvin (`T_0`, ``k_To``)
-        R: the universal gas constant (:math:`R`, ``k_R``)
-        a: intercept of the entropy factor (:math:`a`, ``kattge_knorr_a_ent``)
-        b: slope of the entropy factor (:math:`b`, ``kattge_knorr_b_ent``)
-
-    Returns:
-        Values for :math:`f`
-
-    Examples:
-        >>> # Relative change in Vcmax going (instantaneously, i.e. not
-        >>> # not acclimatedly) from 10 to 25 degrees (percent change):
-        >>> val = ((calc_ftemp_inst_vcmax(25)/calc_ftemp_inst_vcmax(10)-1) * 100)
-        >>> round(val, 4)
-        283.1775
-    """
-
-    # Convert temperatures to Kelvin
-    tkref = pmodel_params.k_To + pmodel_params.k_CtoK
-    tk = tc + pmodel_params.k_CtoK
-
-    # Calculate entropy following Kattge & Knorr (2007): slope and intercept
-    # are defined using temperature in °C, not K!!! 'tcgrowth' corresponds
-    # to 'tmean' in Nicks, 'tc25' is 'to' in Nick's
-    dent = pmodel_params.kattge_knorr_a_ent + pmodel_params.kattge_knorr_b_ent * tc
-    fva = calc_ftemp_arrh(tk, pmodel_params.kattge_knorr_Ha)
-    fvb = (
-        1
-        + np.exp(
-            (tkref * dent - pmodel_params.kattge_knorr_Hd) / (pmodel_params.k_R * tkref)
-        )
-    ) / (
-        1
-        + np.exp((tk * dent - pmodel_params.kattge_knorr_Hd) / (pmodel_params.k_R * tk))
-    )
-
-    return fva * fvb
-
-
-def calc_ftemp_kphio(
-    tc: NDArray, c4: bool = False, pmodel_params: PModelParams = PModelParams()
-) -> NDArray:
-    r"""Calculate temperature dependence of quantum yield efficiency.
-
-    Calculates the temperature dependence of the quantum yield efficiency, as a
-    quadratic function of temperature (:math:`T`). The values of the coefficients depend
-    on whether C3 or C4 photosynthesis is being modelled
-
-    .. math::
-
-        \phi(T) = a + b T - c T^2
-
-    The factor :math:`\phi(T)` is to be multiplied with leaf absorptance and the
-    fraction of absorbed light that reaches photosystem II. In the P-model these
-    additional factors are lumped into a single apparent quantum yield efficiency
-    parameter (argument `kphio` to the class :class:`~pyrealm.pmodel.PModel`).
-
-    Args:
-        tc: Temperature, relevant for photosynthesis (°C)
-        c4: Boolean specifying whether fitted temperature response for C4 plants
-            is used. Defaults to ``False`` to estimate :math:`\phi(T)` for C3 plants.
-        pmodel_params: An instance of :class:`~pyrealm.param_classes.PModelParams`.
-
-    PModel Parameters:
-        C3: the parameters (:math:`a,b,c`, ``kphio_C3``) are taken from the
-            temperature dependence of the maximum quantum yield of photosystem
-            II in light-adapted tobacco leaves determined by :cite:`Bernacchi:2003dc`.
-        C4: the parameters (:math:`a,b,c`, ``kphio_C4``) are taken from
-            :cite:`cai:2020a`.
-
-    Returns:
-        Values for :math:`\phi(T)`
-
-    Examples:
-        >>> # Relative change in the quantum yield efficiency between 5 and 25
-        >>> # degrees celsius (percent change):
-        >>> val = (calc_ftemp_kphio(25.0) / calc_ftemp_kphio(5.0) - 1) * 100
-        >>> round(val, 5)
-        52.03969
-        >>> # Relative change in the quantum yield efficiency between 5 and 25
-        >>> # degrees celsius (percent change) for a C4 plant:
-        >>> val = (calc_ftemp_kphio(25.0, c4=True) /
-        ...        calc_ftemp_kphio(5.0, c4=True) - 1) * 100
-        >>> round(val, 5)
-        432.25806
-    """
-
-    if c4:
-        coef = pmodel_params.kphio_C4
-    else:
-        coef = pmodel_params.kphio_C3
-
-    ftemp = coef[0] + coef[1] * tc + coef[2] * tc**2
-    ftemp = np.clip(ftemp, 0.0, None)
-
-    return ftemp
-
-
-def calc_gammastar(
-    tc: NDArray, patm: NDArray, pmodel_params: PModelParams = PModelParams()
-) -> NDArray:
-    r"""Calculate the photorespiratory CO2 compensation point.
-
-    Calculates the photorespiratory **CO2 compensation point** in absence of dark
-    respiration (:math:`\Gamma^{*}`, ::cite:`Farquhar:1980ft`) as:
-
-    .. math::
-
-        \Gamma^{*} = \Gamma^{*}_{0} \cdot \frac{p}{p_0} \cdot f(T, H_a)
-
-    where :math:`f(T, H_a)` modifies the activation energy to the the local temperature
-    following an Arrhenius-type temperature response function implemented in
-    :func:`calc_ftemp_arrh`. Estimates of :math:`\Gamma^{*}_{0}` and :math:`H_a` are
-    taken from :cite:`Bernacchi:2001kg`.
-
-    Args:
-        tc: Temperature relevant for photosynthesis (:math:`T`, °C)
-        patm: Atmospheric pressure (:math:`p`, Pascals)
-        pmodel_params: An instance of :class:`~pyrealm.param_classes.PModelParams`.
-
-    PModel Parameters:
-        To: the standard reference temperature (:math:`T_0`. ``k_To``)
-        Po: the standard pressure (:math:`p_0`, ``k_Po`` )
-        gs_0: the reference value of :math:`\Gamma^{*}` at standard temperature
-            (:math:`T_0`) and pressure (:math:`P_0`)  (:math:`\Gamma^{*}_{0}`,
-            ``bernacchi_gs25_0``)
-        ha: the activation energy (:math:`\Delta H_a`, ``bernacchi_dha``)
-
-    Returns:
-        A float value or values for :math:`\Gamma^{*}` (in Pa)
-
-    Examples:
-        >>> # CO2 compensation point at 20 degrees Celsius and standard
-        >>> # atmosphere (in Pa) >>> round(calc_gammastar(20, 101325), 5)
-        3.33925
-    """
-
-    # check inputs, return shape not used
-    _ = check_input_shapes(tc, patm)
-
-    return (
-        pmodel_params.bernacchi_gs25_0
-        * patm
-        / pmodel_params.k_Po
-        * calc_ftemp_arrh((tc + pmodel_params.k_CtoK), ha=pmodel_params.bernacchi_dha)
-    )
-
-
-def calc_ns_star(
-    tc: NDArray, patm: NDArray, pmodel_params: PModelParams = PModelParams()
-) -> NDArray:
-    r"""Calculate the relative viscosity of water.
-
-    Calculates the relative viscosity of water (:math:`\eta^*`), given the standard
-    temperature and pressure, using :func:`~pyrealm.pmodel.calc_viscosity_h20`
-    (:math:`v(t,p)`) as:
-
-    .. math::
-
-        \eta^* = \frac{v(t,p)}{v(t_0,p_0)}
-
-    Args:
-        tc: Temperature, relevant for photosynthesis (:math:`T`, °C)
-        patm: Atmospheric pressure (:math:`p`, Pa)
-        pmodel_params: An instance of :class:`~pyrealm.param_classes.PModelParams`.
-
-    PModel Parameters:
-        To: standard temperature (:math:`t0`, ``k_To``)
-        Po: standard pressure (:math:`p_0`, ``k_Po``)
-
-    Returns:
-        A numeric value for :math:`\eta^*` (a unitless ratio)
-
-    Examples:
-        >>> # Relative viscosity at 20 degrees Celsius and standard
-        >>> # atmosphere (in Pa):
-        >>> round(calc_ns_star(20, 101325), 5)
-        1.12536
-    """
-
-    visc_env = calc_viscosity_h2o(tc, patm, pmodel_params=pmodel_params)
-    visc_std = calc_viscosity_h2o(
-        np.array(pmodel_params.k_To),
-        np.array(pmodel_params.k_Po),
-        pmodel_params=pmodel_params,
-    )
-
-    return visc_env / visc_std
-
-
-def calc_kmm(
-    tc: NDArray, patm: NDArray, pmodel_params: PModelParams = PModelParams()
-) -> NDArray:
-    r"""Calculate the Michaelis Menten coefficient of Rubisco-limited assimilation.
-
-    Calculates the Michaelis Menten coefficient of Rubisco-limited assimilation
-    (:math:`K`, ::cite:`Farquhar:1980ft`) as a function of temperature (:math:`T`) and
-    atmospheric pressure (:math:`p`) as:
-
-      .. math:: K = K_c ( 1 + p_{\ce{O2}} / K_o),
-
-    where, :math:`p_{\ce{O2}} = 0.209476 \cdot p` is the partial pressure of oxygen.
-    :math:`f(T, H_a)` is an Arrhenius-type temperature response of activation energies
-    (:func:`calc_ftemp_arrh`) used to correct Michalis constants at standard temperature
-    for both :math:`\ce{CO2}` and :math:`\ce{O2}` to the local temperature (Table 1,
-    ::cite:`Bernacchi:2001kg`):
-
-      .. math::
-        :nowrap:
-
-        \[
-            \begin{align*}
-                K_c &= K_{c25} \cdot f(T, H_{kc})\\
-                K_o &= K_{o25} \cdot f(T, H_{ko})
-            \end{align*}
-        \]
-
-    .. TODO - why this height? Inconsistent with calc_gammastar which uses P_0
-              for the same conversion for a value in the same table.
-
-    Args:
-        tc: Temperature, relevant for photosynthesis (:math:`T`, °C)
-        patm: Atmospheric pressure (:math:`p`, Pa)
-        pmodel_params: An instance of :class:`~pyrealm.param_classes.PModelParams`.
-
-    PModel Parameters:
-        hac: activation energy for :math:`\ce{CO2}` (:math:`H_{kc}`, ``bernacchi_dhac``)
-        hao:  activation energy for :math:`\ce{O2}` (:math:`\Delta H_{ko}`,
-            ``bernacchi_dhao``)
-        kc25: Michelis constant for :math:`\ce{CO2}` at standard temperature
-            (:math:`K_{c25}`, ``bernacchi_kc25``)
-        ko25: Michelis constant for :math:`\ce{O2}` at standard temperature
-            (:math:`K_{o25}`, ``bernacchi_ko25``)
-
-    Returns:
-        A numeric value for :math:`K` (in Pa)
-
-    Examples:
-        >>> # Michaelis-Menten coefficient at 20 degrees Celsius and standard
-        >>> # atmosphere (in Pa):
-        >>> round(calc_kmm(20, 101325), 5)
-        46.09928
-    """
-
-    # Check inputs, return shape not used
-    _ = check_input_shapes(tc, patm)
-
-    # conversion to Kelvin
-    tk = tc + pmodel_params.k_CtoK
-
-    kc = pmodel_params.bernacchi_kc25 * calc_ftemp_arrh(
-        tk, ha=pmodel_params.bernacchi_dhac
-    )
-    ko = pmodel_params.bernacchi_ko25 * calc_ftemp_arrh(
-        tk, ha=pmodel_params.bernacchi_dhao
-    )
-
-    # O2 partial pressure
-    po = pmodel_params.k_co * 1e-6 * patm
-
-    return kc * (1.0 + po / ko)
-
-
-def calc_kp_c4(
-    tc: NDArray, patm: NDArray, pmodel_params: PModelParams = PModelParams()
-) -> NDArray:
-    r"""Calculate the Michaelis Menten coefficient of PEPc.
-
-    Calculates the Michaelis Menten coefficient of phosphoenolpyruvate carboxylase
-    (PEPc) (:math:`K`, :cite:`boyd:2015a`) as a function of temperature (:math:`T`) and
-    atmospheric pressure (:math:`p`) as:
-
-    Args:
-        tc: Temperature, relevant for photosynthesis (:math:`T`, °C)
-        patm: Atmospheric pressure (:math:`p`, Pa)
-        pmodel_params: An instance of :class:`~pyrealm.param_classes.PModelParams`.
-
-    PModel Parameters:
-        hac: activation energy for :math:`\ce{CO2}` (:math:`H_{kc}`,
-             ``boyd_dhac_c4``)
-        kc25: Michelis constant for :math:`\ce{CO2}` at standard temperature
-            (:math:`K_{c25}`, ``boyd_kp25_c4``)
-
-    Returns:
-        A numeric value for :math:`K` (in Pa)
-
-    Examples:
-        >>> # Michaelis-Menten coefficient at 20 degrees Celsius and standard
-        >>> # atmosphere (in Pa):
-        >>> round(calc_kp_c4(20, 101325), 5)
-        9.26352
-    """
-
-    # Check inputs, return shape not used
-    _ = check_input_shapes(tc, patm)
-
-    # conversion to Kelvin
-    tk = tc + pmodel_params.k_CtoK
-    return pmodel_params.boyd_kp25_c4 * calc_ftemp_arrh(
-        tk, ha=pmodel_params.boyd_dhac_c4
-    )
-
-
-def calc_soilmstress(
-    soilm: NDArray,
-    meanalpha: NDArray = np.array(1.0),
-    pmodel_params: PModelParams = PModelParams(),
-) -> NDArray:
-    r"""Calculate Stocker's empirical soil moisture stress factor.
-
-    Calculates an **empirical soil moisture stress factor**  (:math:`\beta`,
-    ::cite:`Stocker:2020dh`) as a function of relative soil moisture (:math:`m_s`,
-    fraction of field capacity) and average aridity, quantified by the local annual mean
-    ratio of actual over potential evapotranspiration (:math:`\bar{\alpha}`).
-
-    The value of :math:`\beta` is defined relative to two soil moisture thresholds
-    (:math:`\theta_0, \theta^{*}`) as:
-
-      .. math::
-        :nowrap:
-
-        \[
-            \beta =
-                \begin{cases}
-                    0
-                    q(m_s - \theta^{*})^2 + 1,  & \theta_0 < m_s <= \theta^{*} \\
-                    1, &  \theta^{*} < m_s,
-                \end{cases}
-        \]
-
-    where :math:`q` is an aridity sensitivity parameter setting the stress factor at
-    :math:`\theta_0`:
-
-    .. math:: q=(1 - (a + b \bar{\alpha}))/(\theta^{*} - \theta_{0})^2
-
-    Default parameters of :math:`a=0` and :math:`b=0.7330` are as described in Table 1
-    of :cite:`Stocker:2020dh` specifically for the 'FULL' use case, with
-    ``method_jmaxlim="wang17"``, ``do_ftemp_kphio=TRUE``.
-
-    Args:
-        soilm: Relative soil moisture as a fraction of field capacity
-            (unitless). Defaults to 1.0 (no soil moisture stress).
-        meanalpha: Local annual mean ratio of actual over potential
-            evapotranspiration, measure for average aridity. Defaults to 1.0.
-        pmodel_params: An instance of :class:`~pyrealm.param_classes.PModelParams`.
-
-    PModel Parameters:
-        theta0: lower bound of soil moisture
-            (:math:`\theta_0`, ``soilmstress_theta0``).
-        thetastar: upper bound of soil moisture
-            (:math:`\theta^{*}`, ``soilmstress_thetastar``).
-        a: aridity parameter (:math:`a`, ``soilmstress_a``).
-        b: aridity parameter (:math:`b`, ``soilmstress_b``).
-
-    Returns:
-        A numeric value or values for :math:`\beta`
-
-    Examples:
-        >>> # Relative reduction (%) in GPP due to soil moisture stress at
-        >>> # relative soil water content ('soilm') of 0.2:
-        >>> round((calc_soilmstress(0.2) - 1) * 100, 5)
-        -11.86667
-    """
-
-    # TODO - presumably this should also have beta(theta) = 0 when m_s <=
-    #        theta_0. Actually, no - these limits aren't correct. This is only
-    #        true when meanalpha=0, otherwise beta > 0 when m_s < theta_0.
-    # TODO - move soilm params into standalone param class for this function -
-    #        keep the PModelParams cleaner?
-
-    # Check inputs, return shape not used
-    _ = check_input_shapes(soilm, meanalpha)
-
-    # Calculate outstress
-    y0 = pmodel_params.soilmstress_a + pmodel_params.soilmstress_b * meanalpha
-    beta = (1.0 - y0) / (
-        pmodel_params.soilmstress_theta0 - pmodel_params.soilmstress_thetastar
-    ) ** 2
-    outstress = 1.0 - beta * (soilm - pmodel_params.soilmstress_thetastar) ** 2
-
-    # Filter wrt to thetastar
-    outstress = np.where(soilm <= pmodel_params.soilmstress_thetastar, outstress, 1.0)
-
-    # Clip
-    outstress = np.clip(outstress, 0.0, 1.0)
-
-    return outstress
-
-
-def calc_viscosity_h2o(
-    tc: NDArray,
-    patm: NDArray,
-    pmodel_params: PModelParams = PModelParams(),
-    simple: bool = False,
-) -> NDArray:
-    r"""Calculate the viscosity of water.
-
-    Calculates the viscosity of water (:math:`\eta`) as a function of temperature and
-    atmospheric pressure (::cite:`Huber:2009fy`).
-
-    Args:
-        tc: air temperature (°C)
-        patm: atmospheric pressure (Pa)
-        pmodel_params: An instance of :class:`~pyrealm.param_classes.PModelParams`.
-        simple: Use the simple formulation.
-
-    Returns:
-        A float giving the viscosity of water (mu, Pa s)
-
-    Examples:
-        >>> # Density of water at 20 degrees C and standard atmospheric pressure:
-        >>> round(calc_viscosity_h2o(20, 101325), 7)
-        0.0010016
-    """
-
-    # Check inputs, return shape not used
-    _ = check_input_shapes(tc, patm)
-
-    if simple or pmodel_params.simple_viscosity:
-        # The reference for this is unknown, but is used in some implementations
-        # so is included here to allow intercomparison.
-        return np.exp(-3.719 + 580 / ((tc + 273) - 138))
-
-    # Get the density of water, kg/m^3
-    rho = calc_density_h2o(tc, patm, pmodel_params=pmodel_params)
-
-    # Calculate dimensionless parameters:
-    tbar = (tc + pmodel_params.k_CtoK) / pmodel_params.huber_tk_ast
-    rbar = rho / pmodel_params.huber_rho_ast
-
-    # Calculate mu0 (Eq. 11 & Table 2, Huber et al., 2009):
-    tbar_pow = np.power.outer(tbar, np.arange(0, 4))
-    mu0 = (1e2 * np.sqrt(tbar)) / np.sum(
-        np.array(pmodel_params.huber_H_i) / tbar_pow, axis=-1
-    )
-
-    # Calculate mu1 (Eq. 12 & Table 3, Huber et al., 2009):
-    h_array = np.array(pmodel_params.huber_H_ij)
-    ctbar = (1.0 / tbar) - 1.0
-    row_j, _ = np.indices(h_array.shape)
-    mu1 = h_array * np.power.outer(rbar - 1.0, row_j)
-    mu1 = np.power.outer(ctbar, np.arange(0, 6)) * np.sum(mu1, axis=(-2))
-    mu1 = np.exp(rbar * mu1.sum(axis=-1))
-
-    # Calculate mu_bar (Eq. 2, Huber et al., 2009), assumes mu2 = 1
-    mu_bar = mu0 * mu1
-
-    # Calculate mu (Eq. 1, Huber et al., 2009)
-    return mu_bar * pmodel_params.huber_mu_ast  # Pa s
-
-
-def calc_patm(elv: NDArray, pmodel_params: PModelParams = PModelParams()) -> NDArray:
-    r"""Calculate atmospheric pressure from elevation.
-
-    Calculates atmospheric pressure as a function of elevation with reference to the
-    standard atmosphere.  The elevation-dependence of atmospheric pressure is computed
-    by assuming a linear decrease in temperature with elevation and a mean adiabatic
-    lapse rate (Eqn 3, ::cite:`BerberanSantos:2009bk`):
-
-    .. math::
-
-        p(z) = p_0 ( 1 - L z / K_0) ^{ G M / (R L) },
-
-    Args:
-        elv: Elevation above sea-level (:math:`z`, metres above sea level.)
-        pmodel_params: An instance of :class:`~pyrealm.param_classes.PModelParams`.
-
-    PModel Parameters:
-        G: gravity constant (:math:`g`, ``k_G``)
-        Po: standard atmospheric pressure at sea level (:math:`p_0`, ``k_Po``)
-        L: adiabatic temperature lapse rate (:math:`L`, ``k_L``),
-        M: molecular weight for dry air (:math:`M`, ``k_Ma``),
-        R: universal gas constant (:math:`R`, `k_R``)
-        Ko: reference temperature in Kelvin (:math:`K_0`, ``k_To``).
-
-    Returns:
-        A numeric value for :math:`p` in Pascals.
-
-    Examples:
-        >>> # Standard atmospheric pressure, in Pa, corrected for 1000 m.a.s.l.
-        >>> round(calc_patm(1000), 2)
-        90241.54
-    """
-
-    # Convert elevation to pressure, Pa. This equation uses the base temperature
-    # in Kelvins, while other functions use this constant in the PARAM units of
-    # °C.
-
-    kto = pmodel_params.k_To + pmodel_params.k_CtoK
-
-    return pmodel_params.k_Po * (1.0 - pmodel_params.k_L * elv / kto) ** (
-        pmodel_params.k_G * pmodel_params.k_Ma / (pmodel_params.k_R * pmodel_params.k_L)
-    )
-
-
-def calc_co2_to_ca(co2: NDArray, patm: NDArray) -> NDArray:
-    r"""Convert :math:`\ce{CO2}` ppm to Pa.
-
-    Converts ambient :math:`\ce{CO2}` (:math:`c_a`) in part per million to Pascals,
-    accounting for atmospheric pressure.
-
-    Args
-        co2 (float): atmospheric :math:`\ce{CO2}`, ppm
-        patm (float): atmospheric pressure, Pa
-
-    Returns:
-        Ambient :math:`\ce{CO2}` in units of Pa
-
-    Examples:
-        >>> np.round(calc_co2_to_ca(413.03, 101325), 6)
-        41.850265
-    """
-
-    return 1.0e-6 * co2 * patm  # Pa, atms. CO2
-
 
 # Design notes on PModel (0.3.1 -> 0.4.0)
 # The PModel until 0.3.1 was a single class taking tc etc. as inputs. However
@@ -2116,213 +1387,6 @@ class JmaxLimitation:
         # pass here, but setting explicitly within the method for clarity.
         self.f_v = np.array([1.0])
         self.f_j = np.array([1.0])
-
-
-class CalcCarbonIsotopes:
-    r"""Calculate :math:`\ce{CO2}` isotopic discrimination.
-
-    This class estimates the fractionation of atmospheric CO2 by photosynthetic
-    pathways to calculate the isotopic compositions and discrimination given the
-    predicted optimal chi from a :class:`~pyrealm.pmodel.PModel` instance.
-
-    Discrimination against carbon 13 (:math:`\Delta\ce{^{13}C}`)  is calculated
-    using C3 and C4 pathways specific methods, and then discrimination against
-    carbon 14 is estimated as :math:`\Delta\ce{^{14}C} \approx 2 \times
-    \Delta\ce{^{13}C}` (:cite:`graven:2020a`). For C3 plants,
-    :math:`\Delta\ce{^{13}C}` is calculated both including and excluding
-    photorespiration, but these are assumed to be equal for C4 plants. The class
-    also reports the isotopic composition of leaves and wood.
-
-    Args:
-        pmodel: A :class:`~pyrealm.pmodel.PModel` instance providing the
-            photosynthetic pathway and estimated optimal chi.
-        d13CO2: Atmospheric isotopic ratio for Carbon 13
-            (:math:`\delta\ce{^{13}C}`, permil).
-        D14CO2: Atmospheric isotopic ratio for Carbon 14
-            (:math:`\Delta\ce{^{14}C}`, permil).
-        params: An instance of :class:`~pyrealm.param_classes.IsotopesParams`,
-            parameterizing the calculations.
-    """
-
-    def __init__(
-        self,
-        pmodel: PModel,
-        D14CO2: NDArray,
-        d13CO2: NDArray,
-        params: IsotopesParams = IsotopesParams(),
-    ):
-        # Check inputs are congruent
-        _ = check_input_shapes(pmodel.env.tc, d13CO2, D14CO2)
-
-        self.params: IsotopesParams = params
-        """The IsotopesParams instance used to calculate estimates."""
-        self.shape: tuple = pmodel.shape
-        """Records the common numpy array shape of array inputs."""
-        self.c4: bool = pmodel.c4
-        """Indicates if estimates calculated for C3 or C4 photosynthesis."""
-
-        # Attributes defined by methods below
-        self.Delta13C_simple: NDArray
-        r"""Discrimination against carbon 13 (:math:`\Delta\ce{^{13}C}`, permil)
-        excluding photorespiration."""
-        self.Delta14C: NDArray
-        r"""Discrimination against carbon 13 (:math:`\Delta\ce{^{13}C}`, permil)
-        including photorespiration."""
-        self.Delta13C: NDArray
-        r"""Discrimination against carbon 14 (:math:`\Delta\ce{^{14}C}`, permil)
-        including photorespiration."""
-        self.d13C_leaf: NDArray
-        r"""Isotopic ratio of carbon 13 in leaves
-        (:math:`\delta\ce{^{13}C}`, permil)."""
-        self.d14C_leaf: NDArray
-        r"""Isotopic ratio of carbon 14 in leaves
-        (:math:`\delta\ce{^{14}C}`, permil)."""
-        self.d13C_wood: NDArray
-        r"""Isotopic ratio of carbon 13 in wood (:math:`\delta\ce{^{13}C}`, permil),
-        given a parameterized post-photosynthetic fractionation."""
-
-        # Could store pmodel, d13CO2, D14CO2 in instance, but really not needed
-        # so try and keep this class simple with a minimum of attributes.
-        # TODO: map methods for delta13C to C3 and C4.
-
-        if self.c4:
-            self.calc_c4_discrimination(pmodel)
-        else:
-            self.calc_c3_discrimination(pmodel)
-
-        # 14C discrimination is twice the 13C discrimination (Graven et al. 2020)
-        self.Delta14C = self.Delta13C * 2
-
-        # Isotopic composition of leaf
-        self.d13C_leaf = (d13CO2 - self.Delta13C) / (1 + self.Delta13C / 1000)
-        self.d14C_leaf = (D14CO2 - self.Delta14C) / (1 + self.Delta14C / 1000)
-
-        # Isotopic composition of wood considering post-photosynthetic fractionation:
-        self.d13C_wood = self.d13C_leaf + self.params.frank_postfrac
-
-    def __repr__(self) -> str:
-        """Generates a string representation of a CalcCarbonIsotopes instance."""
-        return f"CalcCarbonIsotopes(shape={self.shape}, method={self.c4})"
-
-    def calc_c4_discrimination(self, pmodel: PModel) -> None:
-        r"""Calculate C4 isotopic discrimination.
-
-        In this method, :math:`\delta\ce{^{13}C}` is calculated from optimal
-        :math:`\chi` using an empirical relationship estimated by
-        :cite:`lavergne:2022a`.
-
-        Examples:
-            >>> ppar = PModelParams(beta_cost_ratio_c4=35)
-            >>> env = PModelEnvironment(tc=20, patm=101325, co2=400,
-            ...                         vpd=1000, pmodel_params=ppar)
-            >>> mod_c4 = PModel(env, method_optchi='c4_no_gamma')
-            >>> mod_c4_delta = CalcCarbonIsotopes(mod_c4, d13CO2= -8.4, D14CO2 = 19.2)
-            >>> round(mod_c4_delta.Delta13C, 4)
-            5.6636
-            >>> round(mod_c4_delta.d13C_leaf, 4)
-            -13.9844
-        """
-
-        # Equation from C3/C4 paper
-        self.Delta13C_simple = (
-            self.params.lavergne_delta13_a
-            + self.params.lavergne_delta13_b * pmodel.optchi.chi
-        )
-        self.Delta13C = self.Delta13C_simple
-
-    def calc_c4_discrimination_vonC(self, pmodel: PModel) -> None:
-        r"""Calculate C4 isotopic discrimination.
-
-        In this method, :math:`\delta\ce{^{13}C}` is calculated from optimal
-        :math:`\chi` following Equation 1 in :cite:`voncaemmerer:2014a`.
-
-        This method is not yet reachable - it needs a method selection argument to
-        switch approaches and check C4 methods are used with C4 pmodels. The method is
-        preserving experimental code provided by Alienor Lavergne. A temperature
-        sensitive correction term is provided in commented code but not used.
-
-        Examples:
-            >>> ppar = PModelParams(beta_cost_ratio_c4=35)
-            >>> env = PModelEnvironment(tc=20, patm=101325, co2=400,
-            ...                         vpd=1000, pmodel_params=ppar)
-            >>> mod_c4 = PModel(env, method_optchi='c4_no_gamma')
-            >>> mod_c4_delta = CalcCarbonIsotopes(mod_c4, d13CO2= -8.4, D14CO2 = 19.2)
-            >>> # round(mod_c4_delta.Delta13C, 4)  # NOT CHECKED 5.2753
-            >>> # round(mod_c4_delta.d13C_leaf, 4)  # NOT CHECKED -13.6036
-        """
-
-        warn("This method is experimental code from Alienor Lavergne")
-
-        # 13C discrimination (‰): von Caemmerer et al. (2014) Eq. 1
-        self.Delta13C_simple = (
-            self.params.farquhar_a
-            + (
-                self.params.vonCaemmerer_b4
-                + (self.params.farquhar_b - self.params.vonCaemmerer_s)
-                * self.params.vonCaemmerer_phi
-                - self.params.farquhar_a
-            )
-            * pmodel.optchi.chi
-        )
-
-        # Equation A5 from von Caemmerer et al. (2014)
-        # b4 = (-9.483 * 1000) / (273 + self.tc) + 23.89 + 2.2
-        # b4 = self.pmodel_params.vonCaemmerer_b4
-
-        self.Delta13C = self.Delta13C_simple
-
-    def calc_c3_discrimination(self, pmodel: PModel) -> None:
-        r"""Calculate C3 isotopic discrimination.
-
-        This method calculates the isotopic discrimination for
-        :math:`\Delta\ce{^{13}C}` both with and without the photorespiratory
-        effect following :cite:`farquhar:1982a`.
-
-        Examples:
-            >>> env = PModelEnvironment(tc=20, patm=101325, co2=400,
-            ...                         vpd=1000, theta=0.4)
-            >>> mod_c3 = PModel(env, method_optchi='lavergne20_c3')
-            >>> mod_c3_delta = CalcCarbonIsotopes(mod_c3, d13CO2= -8.4, D14CO2 = 19.2)
-            >>> round(mod_c3_delta.Delta13C, 4)
-            20.4056
-            >>> round(mod_c3_delta.d13C_leaf, 4)
-            -28.2296
-        """
-
-        # 13C discrimination (permil): Farquhar et al. (1982)
-        # Simple
-        self.Delta13C_simple = (
-            self.params.farquhar_a
-            + (self.params.farquhar_b - self.params.farquhar_a) * pmodel.optchi.chi
-        )
-
-        # with photorespiratory effect:
-        self.Delta13C = (
-            self.params.farquhar_a
-            + (self.params.farquhar_b2 - self.params.farquhar_a) * pmodel.optchi.chi
-            - self.params.farquhar_f * pmodel.env.gammastar / pmodel.env.ca
-        )
-
-    def summarize(self, dp: int = 2) -> None:
-        """Print summary of values estimated in CalcCarbonIsotopes.
-
-        Prints a summary of the variables calculated within an instance of
-        CalcCarbonIsotopes including the mean, range and number of nan values.
-
-        Args:
-            dp: The number of decimal places used in rounding summary stats.
-        """
-
-        attrs = [
-            ("Delta13C_simple", "permil"),  # ‰
-            ("Delta13C", "permil"),
-            ("Delta14C", "permil"),
-            ("d13C_leaf", "permil"),
-            ("d14C_leaf", "permil"),
-            ("d13C_wood", "permil"),
-        ]
-
-        summarize_attrs(self, attrs, dp=dp)
 
 
 class C3C4Competition:
