+++ setup.cfg
@@ -11,6 +11,9 @@ ignore =
 	# Line break before binary operators is ok
 	D105
 	# Magic methods do not need docstrings.
+per-file-ignores =
+	pyrealm/pmodel/__init__.py: F401
+	# Flattening namespace of submodules
 
 max-line-length = 88
 exclude = tests/*
